be-python-flask/pyproject.toml
[tool.poetry]
name = "long_time_no_see"
version = "0.1.0"
description = "sandbox project"
authors = ["ToKra <kratas.tom@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.1"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
behave = "^1.2.6"
requests = "^2.31.0"
python-dotenv = "^1.0"
pytest = "^8.0"

[tool.poetry.dev-dependencies]
ruff = "^0.1.14"
mypy = "^1.8.0"